인벤토리 스크립트 폴더 구조 정리
