+++ ReBloom/Assets/Scripts/Inventory/0. Data.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c636abebe59d00242bd7e5e1d7cf60c3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
