+++ ReBloom/Assets/Scripts/Inventory/0. Data/InventoryItemData.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 1ac6f7757fcac8344983837fc8d9ffba
