+++ ReBloom/Assets/Scripts/Inventory/1. Core.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8d7802a539d11744cb50cf6de1e5953f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
