+++ ReBloom/Assets/Scripts/Inventory/1. Core/GameInventory.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 164c1d96eefe24b419f67d927d85e55c
