+++ ReBloom/Assets/Scripts/Inventory/2. World.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7e875a7bf4933764abf59182c3274ee0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
