+++ ReBloom/Assets/Scripts/Inventory/3. UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c071cc6f5d1d6dd4e83d7f49bcdbb359
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
