+++ ReBloom/Assets/Scripts/Inventory/3. UI/Common.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 255c2630490875e4686fd34d938e8fe2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
