+++ ReBloom/Assets/Scripts/Inventory/3. UI/Common/ItemSlotBase.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 3752283e718a6254d84d717d4c50d81d
