+++ ReBloom/Assets/Scripts/Inventory/3. UI/Common/ItemTooltip.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: fc11b4f694cf6854cae4ac682eb56ffd
