+++ ReBloom/Assets/Scripts/Inventory/3. UI/Debug.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 44d6fd6895cece14692ee9eaefdaafce
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
