+++ ReBloom/Assets/Scripts/Inventory/3. UI/Debug/DebugInventoryInput.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 054f15256ad782e4db33056148847662
