+++ ReBloom/Assets/Scripts/Inventory/3. UI/Debug/DebugInventoryUI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 612846300bf05bd49b01f63c631731ef
