+++ ReBloom/Assets/Scripts/Inventory/3. UI/Debug/DebugItemSlot.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: fe55f191875333244827468bf0a2d2f9
