+++ ReBloom/Assets/Scripts/Inventory/3. UI/Debug/DebugItemTooltip.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 4dd553b1f6962774eaec8b07d57a7b45
