+++ ReBloom/Assets/Scripts/Inventory/3. UI/DragDrop.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 57732f1cb16bd2b4da29a25b8b30bf07
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
