+++ ReBloom/Assets/Scripts/Inventory/3. UI/DragDrop/InventoryItemDragHandler.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 8a10bc6655d9c574586fbd7a34cdc586
