+++ ReBloom/Assets/Scripts/Inventory/3. UI/DragDrop/InventorySlotDropHandler.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0aa7d57384a786d49a2e1b222102c07d
