+++ ReBloom/Assets/Scripts/Inventory/3. UI/DragDrop/WorldDropZone.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0542ffd864d3daf4ca414ce818eebe9f
