+++ ReBloom/Assets/Scripts/Inventory/3. UI/Game.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a7de78a00cc8881448e911170e40cc80
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
