+++ ReBloom/Assets/Scripts/Inventory/3. UI/Game/GameInventorySlot.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c4a16a597216a6e439f7a27d311966db
