+++ ReBloom/Assets/Scripts/Inventory/3. UI/Game/GameInventoryUI.cs
@@ -1,0 +1,16 @@
+using UnityEngine;
+
+public class GameInventoryUI : MonoBehaviour
+{
+    // Start is called once before the first execution of Update after the MonoBehaviour is created
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
