+++ ReBloom/Assets/Scripts/Inventory/3. UI/Game/GameInventoryUI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 246f9d67620202b45aea595b05025c3e
