+++ ReBloom/Assets/Scripts/Inventory/3. UI/Game/QuickSlot.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: b036c4a84f453694ab2c5d17d1107285
