+++ ReBloom/Assets/Scripts/Inventory/3. UI/Game/QuickSlotUI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 4d0049908818862459879dde21195c61
