--- ReBloom/Assets/Scripts/Inventory/Debug.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 44d6fd6895cece14692ee9eaefdaafce
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
