--- ReBloom/Assets/Scripts/Inventory/Debug/DebugInventoryInput.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 054f15256ad782e4db33056148847662
