--- ReBloom/Assets/Scripts/Inventory/Debug/DebugInventoryUI.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 612846300bf05bd49b01f63c631731ef
