--- ReBloom/Assets/Scripts/Inventory/Debug/DebugItemSlot.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: fe55f191875333244827468bf0a2d2f9
