--- ReBloom/Assets/Scripts/Inventory/Debug/DebugItemTooltip.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 4dd553b1f6962774eaec8b07d57a7b45
