+++ ReBloom/Assets/Scripts/Inventory/ItemIconDragHandler.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2e351fe0def62c64d9805d82e25ba483
