--- ReBloom/Assets/Scripts/Inventory/World.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7e875a7bf4933764abf59182c3274ee0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
