--- ReBloom/Assets/Scripts/Inventory/World/WorldDropZone.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 0542ffd864d3daf4ca414ce818eebe9f
